Proposals
This page shows a list of proposals. Hopefully the transferral of data between this and the vote page works. Again, I may need to update the default web3 account to make it work.
In the next patch I may remove the ability to see inactive proposals and may add a search function.